requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
alembic==1.14.0
pydantic==2.10.4
pydantic-settings==2.7.0
python-multipart==0.0.20
google-generativeai==0.8.3
pdfplumber==0.11.4
openpyxl==3.1.5
pandas>=2.0.0
python-dotenv==1.0.1
reportlab==4.2.5
